Refactor: Decompose Gateway God Object (Phase 2.6 - Part 2)
Breaks apart the monolithic `ramses_rf.Gateway` class to resolve severe Single Responsibility Principle (SRP) violations and prepare the architecture for the upcoming Async Event-Driven pipeline.

- Extracted typed configuration into `ramses_rf/config.py` (GatewayConfig).
- Extracted application lifecycle (start, stop, pause, SQLite DB init, and packet cache restoration) into `ApplicationMixin` in `ramses_rf/app.py`.
- Refactored `ramses_rf/gateway.py` into a Facade that inherits the mixin and maintains 100% backward compatibility for the external API.
- Updated internal `pytest` mocking targets to reflect the new `ramses_rf.app` namespace.
- Maintained 100% `mypy --strict` compliance across the package.

Ref: #639 (OSI Decoupling - Phase 2.6, Part 2)
