+++ src/ramses_rf/app.py
@@ -1,0 +1,283 @@
+#!/usr/bin/env python3
+"""RAMSES RF - Application Lifecycle Management."""
+
+from __future__ import annotations
+
+import asyncio
+import logging
+from datetime import UTC, datetime as dt, timedelta as td
+from logging.handlers import QueueListener
+from typing import TYPE_CHECKING, Any, cast
+
+from ramses_tx import (
+    Packet,
+    extract_known_hgi_id,
+    protocol_factory,
+    set_pkt_logging_config,
+)
+from ramses_tx.logger import flush_packet_log
+
+from .const import DONT_CREATE_MESSAGES, HIGH_VOLUME_STATUS_CODES, I_, RP, RQ, W_, Code
+from .messages import Message
+from .schemas import load_schema
+from .state import MessageStore
+
+if TYPE_CHECKING:
+    from ramses_tx import Engine
+    from ramses_tx.dtos import PacketDTO
+
+    from .config import GatewayConfig
+    from .device import Device
+    from .gateway import Gateway
+    from .interfaces import DeviceRegistryInterface, MessageStoreInterface
+    from .system import Evohome
+
+_LOGGER = logging.getLogger(__name__)
+
+
+class ApplicationMixin:
+    """Application lifecycle management for the Gateway."""
+
+    if TYPE_CHECKING:
+
+        @property
+        def config(self) -> GatewayConfig: ...
+        @property
+        def device_registry(self) -> DeviceRegistryInterface: ...
+
+        _engine: Engine
+        _message_store: MessageStoreInterface | None
+        _pkt_log_listener: QueueListener | None
+        _schema: dict[str, Any]
+
+        def add_task(self, task: asyncio.Task[Any]) -> None: ...
+        def clear_message_history(self) -> None: ...
+        async def schema(self) -> dict[str, Any]: ...
+        async def _msg_handler(self, dto: PacketDTO) -> None: ...
+
+    def create_sqlite_message_index(self) -> None:
+        """Initialize the SQLite MessageStore."""
+        self._message_store = MessageStore(disk_path=self.config.database_path)
+
+    async def start(
+        self,
+        /,
+        *,
+        start_discovery: bool = True,
+        cached_packets: dict[str, dict[str, Any] | str] | None = None,
+    ) -> None:
+        """Start the Gateway and Initiate discovery as required."""
+
+        def initiate_discovery(dev_list: list[Device], sys_list: list[Evohome]) -> None:
+            _LOGGER.debug("Engine: Initiating/enabling discovery...")
+            for device in dev_list:
+                device.discovery.start_poller()
+            for system in sys_list:
+                system.discovery.start_poller()
+                for zone in system.zones:
+                    zone.discovery.start_poller()
+                if system.dhw:
+                    system.dhw.discovery.start_poller()
+
+        _, self._pkt_log_listener = await set_pkt_logging_config(
+            cc_console=(self.config.reduce_processing >= DONT_CREATE_MESSAGES),
+            **self._engine._packet_log,
+        )  # type: ignore[arg-type]
+
+        if self._pkt_log_listener:
+            self._pkt_log_listener.start()
+
+            pkt_log_config = cast("dict[str, Any]", self._engine._packet_log)
+            if flush_interval := pkt_log_config.get("flush_interval", 0):
+
+                async def _periodic_flush() -> None:
+                    try:
+                        while True:
+                            await asyncio.sleep(flush_interval)
+                            await self._engine._loop.run_in_executor(
+                                None, flush_packet_log, self._pkt_log_listener
+                            )
+                    except asyncio.CancelledError:
+                        pass
+
+                self.add_task(self._engine._loop.create_task(_periodic_flush()))
+
+        _LOGGER.info("Ramses RF starts central MessageStore")
+        self.create_sqlite_message_index()
+
+        self.config.disable_discovery, disable_discovery = (
+            True,
+            self.config.disable_discovery,
+        )
+
+        load_schema(
+            cast("Gateway", self), known_list=self._engine._include, **self._schema
+        )
+
+        if cached_packets:
+            await self._restore_cached_packets(cached_packets)
+
+        await self._engine.start()
+
+        self.config.disable_discovery = disable_discovery
+
+        if (
+            not self._engine._disable_sending
+            and not self.config.disable_discovery
+            and start_discovery
+        ):
+            initiate_discovery(
+                self.device_registry.devices, self.device_registry.systems
+            )
+
+    async def stop(self) -> None:
+        """Stop the Gateway and tidy up."""
+        self.config.disable_discovery = True
+        await self._engine.stop()
+
+        if self._pkt_log_listener:
+
+            def _stop_listener(listener: QueueListener) -> None:
+                listener.stop()
+                for handler in listener.handlers:
+                    handler.close()
+
+            await self._engine._loop.run_in_executor(
+                None, _stop_listener, self._pkt_log_listener
+            )
+            self._pkt_log_listener = None
+
+        if self._message_store:
+            self._message_store.stop()
+
+    async def _pause(self, *args: Any) -> None:
+        """Pause the (unpaused) gateway (disables sending/discovery)."""
+        _LOGGER.debug("Gateway: Pausing engine...")
+
+        self.config.disable_discovery, disc_flag = (
+            True,
+            self.config.disable_discovery,
+        )
+
+        try:
+            await self._engine._pause(disc_flag, *args)
+        except RuntimeError:
+            self.config.disable_discovery = disc_flag
+            raise
+
+    async def _resume(self) -> tuple[Any, ...]:
+        """Resume the (paused) gateway."""
+        args: tuple[Any, ...]
+        _LOGGER.debug("Gateway: Resuming engine...")
+
+        self.config.disable_discovery, *args = (
+            await self._engine._resume()  # type: ignore[assignment]
+        )
+
+        return tuple(args)
+
+    async def get_state(
+        self, include_expired: bool = False
+    ) -> tuple[dict[str, Any], dict[str, Any]]:
+        """Return the current schema & state (may include expired packets)."""
+        await self._pause()
+
+        def wanted_msg(msg: Message, include_expired: bool = False) -> bool:
+            if msg.code == Code._313F:
+                return msg.verb in (I_, RP)
+            if getattr(msg, "_expired", False) and not include_expired:
+                return False
+            if msg.code == Code._0404:
+                return msg.verb in (I_, W_) and msg.len > 7
+            if msg.verb in (W_, RQ):
+                return False
+            return include_expired or not getattr(msg, "_expired", False)
+
+        pkts: dict[str, Any] = {}
+        if self._message_store:
+            all_msgs = await self._message_store.all(include_expired=True)
+            for i, msg in enumerate(all_msgs):
+                if wanted_msg(msg, include_expired=include_expired):
+                    dtm_str = msg.dtm.isoformat(timespec="microseconds")
+                    pkts[dtm_str] = msg.dto.source_packets[0].__dict__
+                if i > 0 and i % 100 == 0:
+                    await asyncio.sleep(0)
+
+        await self._resume()
+        return await self.schema(), dict(sorted(pkts.items()))
+
+    async def _restore_cached_packets(
+        self, packets: dict[str, dict[str, Any] | str], _clear_state: bool = False
+    ) -> None:
+        """Restore cached packets (may include expired packets)."""
+
+        def clear_state() -> None:
+            _LOGGER.info("Gateway: Clearing existing schema/state...")
+            cast("Gateway", self)._tcs = None
+            self.device_registry.devices.clear()
+            self.device_registry.device_by_id.clear()
+            self.clear_message_history()
+
+        _LOGGER.debug("Gateway: Restoring a cached packet log...")
+        await self._pause()
+
+        if _clear_state:
+            clear_state()
+
+        enforce_include_list = bool(
+            self._engine._enforce_known_list
+            and extract_known_hgi_id(
+                self._engine._include,
+                disable_warnings=True,
+                strict_checking=True,
+            )
+        )
+
+        tmp_protocol = protocol_factory(
+            self._msg_handler,
+            disable_sending=True,
+            enforce_include_list=enforce_include_list,
+            exclude_list=self._engine._exclude,
+            include_list=self._engine._include,
+        )
+
+        cutoff_dtm = dt.now(tz=UTC) - td(hours=1)
+
+        for i, (dtm, state) in enumerate(packets.items()):
+            if i > 0 and i % 100 == 0:
+                await asyncio.sleep(0)
+
+            try:
+                clean_dtm = dtm.replace("Z", "+00:00")
+                pkt_dtm = dt.fromisoformat(clean_dtm)
+                if pkt_dtm.tzinfo is None:
+                    pkt_dtm = pkt_dtm.replace(tzinfo=UTC)
+                is_old = pkt_dtm < cutoff_dtm
+            except (TypeError, ValueError):
+                is_old = False
+
+            if is_old:
+                is_match = False
+                if isinstance(state, dict) and "code" in state:
+                    is_match = state["code"] in HIGH_VOLUME_STATUS_CODES
+                else:
+                    frame_str = (
+                        state.get("frame", "")
+                        if isinstance(state, dict)
+                        else str(state)
+                    )
+                    is_match = any(
+                        f" {c} " in frame_str for c in HIGH_VOLUME_STATUS_CODES
+                    )
+
+                if is_match:
+                    continue
+
+            try:
+                pkt = Packet.from_dict(dtm, state)
+                tmp_protocol.pkt_received(pkt)
+            except Exception as err:
+                _LOGGER.debug("Gateway: Failed to restore packet %s: %s", dtm, err)
+
+        _LOGGER.debug("Gateway: Restored, resuming")
+        await self._resume()
